requirements.txt
a2a-sdk
google-generativeai
python-dotenv
